// src/main/webapp/app/generated/model/created-job-dto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API Version: v0
 * 
 *
 * NOTE: This file is auto-generated. Do not edit manually.
 */


export interface CreatedJobDTO {
    readonly avatar?: string;
    readonly createdAt?: string;
    readonly jobId: string;
    readonly lastModifiedAt?: string;
    readonly professorName?: string;
    readonly startDate?: string;
    readonly state?: CreatedJobDTOStateEnum;
    readonly title: string;
}

export type CreatedJobDTOStateEnum = 'DRAFT' | 'PUBLISHED' | 'CLOSED' | 'APPLICANT_FOUND';

export const CreatedJobDTOStateEnum = {
    Draft: 'DRAFT' as const,
    Published: 'PUBLISHED' as const,
    Closed: 'CLOSED' as const,
    ApplicantFound: 'APPLICANT_FOUND' as const,
} as const;

export const CreatedJobDTOStateEnumValues = ['DRAFT', 'PUBLISHED', 'CLOSED', 'APPLICANT_FOUND'] as const;